后端精度更改

+++ elmboot/src/main/java/com/neusoft/elmboot/po/Food.java
@@ -5,51 +5,64 @@ public class Food {
 	private String foodName;
 	private String foodExplain;
 	private String foodImg;
-	private Double foodPrice;
+	private Integer foodPrice;
 	private Integer businessId;
 	private String remarks;
+
 	public Integer getFoodId() {
 		return foodId;
 	}
+
 	public void setFoodId(Integer foodId) {
 		this.foodId = foodId;
 	}
+
 	public String getFoodName() {
 		return foodName;
 	}
+
 	public void setFoodName(String foodName) {
 		this.foodName = foodName;
 	}
+
 	public String getFoodExplain() {
 		return foodExplain;
 	}
+
 	public void setFoodExplain(String foodExplain) {
 		this.foodExplain = foodExplain;
 	}
+
 	public String getFoodImg() {
 		return foodImg;
 	}
+
 	public void setFoodImg(String foodImg) {
 		this.foodImg = foodImg;
 	}
-	public Double getFoodPrice() {
+
+	public Integer getFoodPrice() {
 		return foodPrice;
 	}
-	public void setFoodPrice(Double foodPrice) {
+
+	public void setFoodPrice(Integer foodPrice) {
 		this.foodPrice = foodPrice;
 	}
+
 	public Integer getBusinessId() {
 		return businessId;
 	}
+
 	public void setBusinessId(Integer businessId) {
 		this.businessId = businessId;
 	}
+
 	public String getRemarks() {
 		return remarks;
 	}
+
 	public void setRemarks(String remarks) {
 		this.remarks = remarks;
 	}
 
 }
-
